Move non-dev imports off of tests.contrib_base_classes
